% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etm_prefit.R
\name{etm_prefit}
\alias{etm_prefit}
\title{Use Pre-Fitted Word Embeddings with the ETM algorithm}
\usage{
etm_prefit(dataset_name, data_path, embed_path, num_topics = 50,
  epochs = 1000)
}
\arguments{
\item{dataset_name}{The name of the dataset.}

\item{data_path}{The directory or location to the dataset.}

\item{embed_path}{The directory or location to the embeddings file.}

\item{num_topics}{The number of topics to learn.  The default number of topics is 50.}

\item{epochs}{Number of times the dataset is fed into the algorithm for
learning.  The default number of epochs is 1000.  Minimum number of epochs is 1.}
}
\description{
`etm_prefit` runs the ETM algorithm on text with pre-fitted word embeddings
(e.g. using skiggrams with the `etm_skipgram` function).
}
\examples{
\dontrun{
etm_prefit(dataset_name = "corpus_1", data_path = "Desktop/", embed_path = "Desktop/embed/",
 num_topics = 50, epochs = 1000)
}
}
