% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etm_preprocess.R
\name{etm_preprocess}
\alias{etm_preprocess}
\title{Preprocess the Corpus}
\usage{
etm_preprocess(data_path, max_df = 0.7, min_df = 100)
}
\arguments{
\item{data_path}{Path to the corpus text file.}

\item{max_df}{The desired value for maximum document frequency.  Default is 0.7.}

\item{min_df}{The desired value for minimum document frequency.  Default is 100.}
}
\value{
A folder with ".mat" files of bag-of-words representation.  The folder will be
located in the same directory as the corpus text file.
}
\description{
`etm_preprocess` preprocesses the corpus to convert the text into bag-of-words
representation (tokens and counts).  `etm_preprocess` is the precursor step before
using word embeddings (`etm_embed`) or modeling the embedded topic model (`etm_main`).
}
\examples{
\dontrun{
etm_preprocess(data_path = "Desktop/corpus.txt")

etm_preprocess(data_path = "Documents/news.txt", max_df = 0.8, min_df = 100)
}
}
