% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_py_package.R
\name{install_py_package}
\alias{install_py_package}
\title{Manually Install Python Packages}
\usage{
install_py_package(packages = c("gensim", "matplotlib", "numpy",
  "sklearn", "scipy", "torch"), install_method = c("conda"),
  conda = "auto", env_name = "r-reticulate", pip = FALSE)
}
\arguments{
\item{packages}{A vector of the package names to install.  By default, this will install all
of the necessary Python packages, namely: gensim, matplotlib, numpy, scikit-learn (sklearn),
scipy and PyTorch (torch). Boolean (TRUE or FALSE) entry, with FALSE as default.
Note that if the installation method is "conda", the PyTorch package is named "pytorch", whereas if the
installation method is "virtualenv", then the PyTorch package is named "torch".}

\item{install_method}{Installation method for the Python packages.  By default,
the installation method is "conda".  Linus and Mac users may also select the installation
method "virtualenv" (Windows users are limited to only the "conda" installation method).}

\item{conda}{The path to a conda executable.  The default option, "auto",
will automatically find an appropriate conda binary (see \code{\link[reticulate]{py_install}} for more information).}

\item{env_name}{The name of the virtual environment in which Python packages are to
be installed.  Users are highly encouraged to keep the virtual environment name as the default name,
"r-reticulate", for consistency.}

\item{pip}{Uses pip (the native python package installer) for package installation.
Boolean entry, with "FALSE" as default.}
}
\description{
By default, ETMr should automatically install all necessary Python packages.
Nevertheless, this function provides a way to manually install all necessary Python
packages.  This function is useful in two settings: first, to install and load Python
packages in a non-default virtual environment (not "r-reticulate"), or second, to
install versions of Python packages to utilize GPUs for machine learning.  "install_py_packages"
provides parameters to install either all Python packages, or just the GPU-optimized packages (PyTorch).
}
\examples{
\dontrun{
install_py_package()
install_py_package(c("pytorch", "torchvision"), install_method = c("conda"))
install_py_package(c("torch", "torchvision"), install_method = c("virtualenv"))
}
}
