% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pytorch_gpu.R
\name{install_pytorch_gpu}
\alias{install_pytorch_gpu}
\title{Install GPU-enabled PyTorch}
\usage{
install_pytorch_gpu(os = "windows")
}
\arguments{
\item{os}{Your operating system, either Windows or Linux (there is no GPU support for
Mac OS users).  Default operating system is Windows.}
}
\description{
`install_pytorch_gpu` installs a version of PyTorch that enables GPUs for general purpose
processing.  GPU support for PyTorch is only available for Windows and Linux users
(not Mac OS users).  `install_pytorch_gpu`  will first uninstall all CPU-based versions of
PyTorch, and then install the GPU-based version of PyTorch.
}
\examples{
\dontrun{
install_pytorch_gpu(os = "windows")
install_pytorch_gpu(os = "linux")
}
}
