% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pytorch_check_gpu.R
\name{pytorch_check_gpu}
\alias{pytorch_check_gpu}
\title{Check that GPU support is enabled for PyTorch}
\usage{
pytorch_check_gpu()
}
\value{
Boolean (TRUE or FALSE) entry, depending on if GPU support
is available.  Note that GPU support is not available for Mac OS users.
}
\description{
`pytorch_check_gpu` checks if GPU support is enabled for PyTorch.
}
\examples{
\dontrun{
pytorch_check_gpu()
}
}
