% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_etm.R
\name{start_etm}
\alias{start_etm}
\title{Start Python Environment}
\usage{
start_etm(env_name = "r-reticulate")
}
\arguments{
\item{env_name}{By default, ETMr automatically loads a virtual environment
within which R and Python packages can be simultaneously loaded.  The default virtual environment name
is "r-reticulate".  Advanced users are welcome to change the virtual environment name;
just remember to re-install all necessary Python packages using the function `install_py_package`.}
}
\description{
This is always the first step when using `ETMr`.  `start_etm`
first launches a virtual environment in which Python can be used, then
verifies that all necessary Python packages are available.
}
\examples{
\dontrun{
start_etm()
start_etm("new_environment_name")
}
}
